Lab4/requirements.txt
gostcrypto
pycryptodome
cryptography
